Fix added integer int, unsignet int

+++ libavcodec/libshine/bitstream.c
@@ -19,7 +19,7 @@ void shine_open_bit_stream(bitstream_t *bs, int size) {
   bs->data = (unsigned char *)malloc(size * sizeof(unsigned char));
   bs->data_size = size;
   bs->data_position = 0;
-  bs->cache = 0;
+  bs->cache = 0u;
   bs->cache_bits = 32;
 }
 
@@ -39,35 +39,35 @@ void shine_close_bit_stream(bitstream_t *bs) {
  */
 void shine_putbits(bitstream_t *bs, unsigned int val, unsigned int N) {
 #ifdef DEBUG
-  if (N > 32)
+  if (N > 32u)
     printf("Cannot write more than 32 bits at a time.\n");
-  if (N < 32 && (val >> N) != 0)
+  if (N < 32u && (val >> N) != 0u)
     printf("Upper bits (higher than %d) are not all zeros.\n", N);
 #endif
 
-  if (bs->cache_bits > N) {
-    bs->cache_bits -= N;
+  if (bs->cache_bits > (int)N) {
+    bs->cache_bits -= (int)N;
     bs->cache |= val << bs->cache_bits;
   } else {
-    if (bs->data_position + sizeof(unsigned int) >= bs->data_size) {
+    if (bs->data_position + (int)sizeof(unsigned int) >= bs->data_size) {
       bs->data = (unsigned char *)realloc(bs->data,
                                           bs->data_size + (bs->data_size / 2));
       bs->data_size += (bs->data_size / 2);
     }
 
-    N -= bs->cache_bits;
-    bs->cache |= val >> N;
+    N -= (unsigned int)bs->cache_bits;
+    bs->cache |= val >> (int)N;
 #ifdef SHINE_BIG_ENDIAN
     *(unsigned int *)(bs->data + bs->data_position) = bs->cache;
 #else
     *(unsigned int *)(bs->data + bs->data_position) = SWAB32(bs->cache);
 #endif
-    bs->data_position += sizeof(unsigned int);
-    bs->cache_bits = 32 - N;
-    if (N != 0)
+    bs->data_position += (int)sizeof(unsigned int);
+    bs->cache_bits = 32 - (int)N;
+    if (N != 0u)
       bs->cache = val << bs->cache_bits;
     else
-      bs->cache = 0;
+      bs->cache = 0u;
   }
 }
